# Conflicts:
#	cadicom/src/main/java/es/uca/cadicom/entity/Usuario.java
#	cadicom/src/main/java/es/uca/cadicom/views/frontoffice/ConsumoView.java

+++ cadicom/src/main/java/es/uca/cadicom/repository/TarifaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TarifaRepository extends JpaRepository<Tarifa, Integer> {
+public interface TarifaRepository extends JpaRepository<Tarifa, Long> {
 }
